refactor: split command substitution
