+++ src/cmd_subst.h
@@ -1,0 +1,8 @@
+#ifndef CMD_SUBST_H
+#define CMD_SUBST_H
+
+char *command_output(const char *cmd);
+char *parse_substitution(char **p);
+
+#endif /* CMD_SUBST_H */
+
